refactor: split `getOpenAiResponse` and `getAnthropicResponse` methods into dedicated files
