+++ apps/hash-ai-worker-ts/src/activities/shared/get-llm-response/max-retry-count.ts
@@ -1,0 +1,4 @@
+/**
+ * The maximum number of times to retry a request to the LLM API.
+ */
+export const maxRetryCount = 3;
